requirements.txt
# NOC Toolkit - Python Dependencies
# =====================================
# This file contains all dependencies required by the toolkit and its tools

# Core dependencies
# -----------------
python-dotenv>=1.0.0  # Environment variable management
PyYAML>=6.0           # YAML config file parsing

# PagerDuty integration
# ---------------------
pagerduty>=1.0.0  # PagerDuty Python REST API client

# Jira integration
# ----------------
jira>=3.5.0  # Jira Python library for REST API

# User interface
# --------------
tqdm>=4.65.0  # Progress bars for visual feedback
customtkinter>=5.2.0  # Modern themed Tkinter widgets for GUI launcher
Pillow>=10.0.0  # Image support (customtkinter dependency)

# Excel report manipulation
# ------------------------
openpyxl>=3.1.0  # Read/write Excel files for NOC Report Assistant

# Testing
# -------
freezegun>=1.2.0  # Time-freezing for deterministic datetime tests
mutmut>=2.5.0  # Mutation testing for test quality verification

# Optional enhancements
# ---------------------
# colorama>=0.4.6  # Cross-platform colored terminal text (future enhancement)
# tabulate>=0.9.0  # Pretty-print tabular data (future enhancement)
